-se pudo eliminar un area de geofence de la base de datos y del mapa

+++ Android/AbuMonitor/mobile/src/main/java/com/example/comunicationwearmobile/ui/viewmodel/ViewmodelAreaGeofence.kt
@@ -81,7 +81,7 @@ class ViewmodelAreaGeofence(application: Application):AndroidViewModel(applicati
         }
     }
 
-    fun deleteAreaGeofence(idArea:Int){
+    fun deleteAreaGeofence(idArea:Long){
         var rowEliminated:Int
         try {
             viewModelScope.launch {
